+ Reinstaurado el color como RGB aunque se deja la posibilidad del string. + Nueva clase aux_structs.h que se encarga de mantener cosas de estructura compartida. (Donde se encuentra Color)

+++ Muphic/common/src/Figura.cpp
@@ -70,7 +70,7 @@ float Figura::getColorDifference()
 	return colorDifference;
 }
 //------Setters------//
-void Figura::setRGB(float r, float g, float b)
+void Figura::setRGB(int r, int g, int b)
 {
 	rgb.r = r;
 	rgb.g = g;
